chore(engine): move clients in new package
